Move sort interpreters function
